config/requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
pymongo==4.6.1

# AI and ML
langchain==0.1.9
openai==1.12.0
google-generativeai==0.3.2
transformers==4.37.2
torch==2.2.0
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0

# Web and API
requests==2.31.0
aiohttp==3.9.3
beautifulsoup4==4.12.3
selenium==4.18.1
playwright==1.41.2

# Document processing
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Development tools
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Utilities
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
aiofiles==23.2.1 